% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AirPollWeekly-data.R
\docType{data}
\name{AirPollWeekly}
\alias{AirPollWeekly}
\title{Weekly average NO2 and PM2.5 data}
\format{An object of class \code{"data.frame"}.}
\usage{
data(AirPollWeekly)
}
\description{
Weekly average NO2 and PM2.5  data for 1000 hypothetical births.
}
\details{
PM2.5 data has column names PM_1, ..., PM_37 and NO2 data has column names NO2_1, ..., NO2_37.
}
\examples{
data(AirPollWeekly)
matplot(t(AirPollWeekly[1:5,-c(1:2)]),type="l")
}
\keyword{datasets}

