% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdlim.R
\name{bdlim}
\alias{bdlim}
\title{Bayesian distributed lag interaction models}
\usage{
bdlim(Y, X, Z, G = NULL, inter.model = "all", family = gaussian,
  niter = 1000, nburn = round(niter/2), nthin = 1, basis.opts, prior,
  seed)
}
\arguments{
\item{Y}{Vector of outcomes.}

\item{X}{Matrix of exposures over an evenly spaced grid. There should be no missing exposures.}

\item{Z}{Design matrix of covariates. The design matrix for G will be added automatically, including an intercept.}

\item{G}{Vector indicating group membership. This should be a factor.}

\item{inter.model}{Indicator of how G enters the model. If 'bw' both the betas and weight functions will vary between groups. If 'b' then only the betas will vary between groups. If 'w' then only the wieght function will vary by group. If "n" then neither beta or the wieght function will vary by group. If a vector of any combination of the four choices is provided then those models will be run and compared. The option 'all' runs all four models and compares them.}

\item{family}{A description of the error distribution and link function to be used in the model. Currently, gaussian and binomial are supported.}

\item{niter}{Number of MCMC iterations including burnin.}

\item{nburn}{Number of MCMC iterations to be discarded as burnin.}

\item{nthin}{Number of draws taken to obtain one sample.}

\item{basis.opts}{List with the entries: type = the type of basis used, either 'face' (default) or "ns" or "bs" for splines or "gam" for presmoothing the exposure with a gam following defaults from mgcv; knots = the number of knots used for method face; pve = the percent of variance explained by the PCs for method face; df = the df for ns method.}

\item{prior}{List with the entries: sigma = a numeric 2-vector with the shape and rate paramters for the pirior in the error precision (1/sigma^2); betavar = the prior variance for beta; and gamma = the prior variance for the covarites. The priors on beta and gamma are iid normal mean zero.}

\item{seed}{A seed to be set before each model is run.}
}
\value{
An object of class 'bdlim'.
}
\description{
This estimates the Bayesian distributed lag interaction model (BDLIM).
}
\author{
Ander Wilson
}
