% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdlimbasis.R
\name{bdlimbasis}
\alias{bdlimbasis}
\title{Build BDLIM basis}
\usage{
bdlimbasis(X, basis.opts)
}
\arguments{
\item{X}{A n x p matrix of observations. Each row has the observations for one individual. The observations should be evenly spaced over a grid.}

\item{basis.opts}{List with the entries: type = the type of basis used, either 'face' (default) or "ns" or "bs" for splines or "gam" for presmoothing the exposure with a gam following defaults from mgcv; knots = the number of knots used for method face; pve = the percent of variance explained by the PCs for method face; df = the df for ns method.}
}
\description{
\code{bdlimbasis} builds the basis for thge BDLIM model using either FACE of natural cubic splines.
}
\seealso{
\code{\link{fpca.face}}, \code{\link{ns}}, \code{\link{bs}}, and \code{\link{gam}} for details on the methods called.
}
\author{
Ander Wilson
}
