% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bdlimglmall.R
\name{bdlimglmall}
\alias{bdlimglmall}
\title{BDLIM for glm}
\usage{
bdlimglmall(Y, X, Z, G, B, model, niter, nburn, nthin, prior, post)
}
\arguments{
\item{Y}{Outcome vector}

\item{X}{Exposure matrix}

\item{Z}{Matrix of covariates. An intercept will be added}

\item{G}{Vector indicating group membership. This should be a factor.}

\item{B}{Basis object from build.basis}

\item{model}{The inter.model.}

\item{niter}{Number of MCMC iterations}

\item{nburn}{Number of MCMC iterations to be discarded as burning}

\item{nthin}{Number of draws taken to obtainone sampl}

\item{prior}{A vector of lenght 2 whose elements (a,b) are the hyper parameters of a Gamma prior on sigma^(-2) with mean a/b.}

\item{post}{Relating to the likelihood functions.}
}
\description{
This estimates the model for a single group or the overall effect
}
\author{
Ander Wilson
}

