% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bdlimlmb.R
\name{bdlimlmb}
\alias{bdlimlmb}
\title{BDLIM-b for linear model}
\usage{
bdlimlmb(Y, X, Z, G, B, niter, nburn, nthin, prior)
}
\arguments{
\item{X}{Exposure matrix}

\item{Z}{Matrix of covariates. An intercept will be added}

\item{G}{Vector indicating group membership}

\item{B}{Basis object from build.basis}

\item{niter}{Number of MCMC iterations}

\item{nburn}{Number of MCMC iterations to be discarded as burning}

\item{nthin}{Number of draws taken to obtainone sampl}

\item{y}{Outcome vector}

\item{prior.sigma}{A vector of lenght 2 whose elements (a,b) are the hyper parameters of a Gamma prior on sigma^(-2) with mean a/b.}
}
\description{
This estimates the model for a common weight funciton but group specific effects
}
\author{
Ander Wilson
}

