% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bdlimlmbw.R
\name{bdlimlmbw}
\alias{bdlimlmbw}
\title{BDLIM-bw for linear model}
\usage{
bdlimlmbw(Y, X, Z, G, B, niter, nburn, nthin, prior)
}
\arguments{
\item{X}{Exposure matrix}

\item{Z}{Matrix of covariates. An intercept will be added}

\item{G}{Vector indicating group membership}

\item{B}{Basis object from build.basis}

\item{niter}{Number of MCMC iterations}

\item{nburn}{Number of MCMC iterations to be discarded as burning}

\item{nthin}{Number of draws taken to obtainone sampl}

\item{prior}{A vector of lenght 2 whose elements (a,b) are the hyper parameters of a Gamma prior on sigma^(-2) with mean a/b.}

\item{y}{Outcome vector}
}
\description{
This estimates the model for a group specific weight funciton and group specific effects
}
\author{
Ander Wilson
}

