% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bdlimlmoverall.R
\name{bdlimlmoverall}
\alias{bdlimlmoverall}
\title{bdlimlmoverall}
\usage{
bdlimlmoverall(Y, X, Z, G, B, niter, nburn, nthin, prior)
}
\arguments{
\item{X}{Exposure matrix}

\item{Z}{Matrix of covariates. An intercept will be added}

\item{B}{Basis object from build.basis}

\item{niter}{Number of MCMC iterations}

\item{nburn}{Number of MCMC iterations to be discarded as burning}

\item{nthin}{Number of draws taken to obtainone sampl}

\item{prior}{A vector of lenght 2 whose elements (a,b) are the hyper parameters of a Gamma prior on sigma^(-2) with mean a/b.}

\item{y}{Outcome vector}
}
\description{
This estimates the model for a single group or the overall effect
}
\author{
Ander Wilson
}

