% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bkmrdlm.R
\name{bkmrdlm}
\alias{bkmrdlm}
\title{Bayesian kernel machine regression - distributed lag model}
\usage{
bkmrdlm(y, x, z, niter, nburn = round(niter/2), nthin = 1,
  prior_sigma = c(0.1, 0.1), prior_tau = 0.1, kappa = 1,
  basis.opts = list(type = "gam", pve = 0.9), shrink = FALSE,
  gaussian = TRUE, polydegree = 2)
}
\arguments{
\item{y}{Vector of outcomes.}

\item{x}{A list of matricies. Each Matric should be n by T with each row being the T-vector of exposures for one individuals. When only one exposure is used x can be a matrix instead of a list.}

\item{z}{a matrix or design frame of covaiates and confounds. This can be ommited.}

\item{niter}{Number of MCMC iterations including burnin.}

\item{nburn}{The number of iterations to be discarded as burnin.}

\item{nthin}{Thining, every nthin-th draw from the posterior will be saved.}

\item{prior_sigma}{Vector of length 2 with the parameters for the gamma prior on sigma^{-2}}

\item{prior_tau}{the prior variance on log(tau^2).}

\item{kappa}{scale parameter, rho/kappa~chisq(1).}

\item{basis.opts}{List with the entries: type = the type of basis used, either 'face' (default) or "ns" or "bs" for splines or "gam" for presmoothing the exposure with a gam following defaults from mgcv; knots = the number of knots used for method face; pve = the percent of variance explained by the PCs for method face; df = the df for ns method.}

\item{shrink}{use the shrinkage prior on theta (default FALSE)}

\item{gaussian}{Use a Gaussian kernel (TRUE, default) or a polynomial kernel (FALSE)}

\item{polydegree}{Degree of polynomial when polynomial kernel is used.  Only applies when gaussian=FALSE.}
}
\value{
An object of class 'bkmrdlm'.
}
\description{
This estimates the Bayesian kernel machine regression - distributed lag model (BKMR-DLM).
}
\author{
Ander Wilson
}
