% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bkmrdlm_multi}
\alias{bkmrdlm_multi}
\title{Estimates the BKMR-DLM for a multiple time-varying predictor.}
\usage{
bkmrdlm_multi(yz, Xlist, b1, a1, a2, kappa, n_inner, n_outer, n_burn)
}
\arguments{
\item{yz}{a matrix that is cbind(y,Z) where Z is a matrix of covariates that does not include an intercept and y is the vector of outcomes.}

\item{Xlist}{a list of matrices each for a single exposure in time order.}

\item{b1}{first parameter for prior on tau^{-2} in the text.}

\item{a1}{first parameter for prior on sigma^{-2}.}

\item{a2}{second parameter for prior on sigma^{-2}.}

\item{kappa}{scale parameter, rho/kappa~chisq(1).}

\item{n_inner}{number of MCMC iterations to run in the inner loop.  This is equivelent the the thinning number.  n_outer*n_inner iteraction will be run and n_outer iterations will be saved.}

\item{n_outer}{number of MCMC iteration in the outer loop.   The output for each is saved.}

\item{n_burn}{number of MCMC iteration to be discarded as burn-in.}
}
\description{
Estimates the BKMR-DLM for a multiple time-varying predictor.
}
