% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlm.R
\name{dlm}
\alias{dlm}
\title{Distributed lag model}
\usage{
dlm(Y, X, Z, df, family = gaussian)
}
\arguments{
\item{Y}{Vector of outcomes.}

\item{X}{Matrix of exposures over an evenly spaced grid. There should be no missing exposures.}

\item{Z}{Optional design matrix or data.frame of covariates. An intercept will be added}

\item{df}{Degrees of freedom for the natrual spline basis}

\item{family}{A description of the error distribution and link function to be used in the model.}
}
\value{
An object of class dlm.  This is a list containing: ``dlm`` the fitted distribued lag function; ``cumulative`` estimates of the cumulative effect; and ``modelfit`` which contains AIC and BIC for the fitted model; ``basis`` is the natrual spline basis for the dlm; and ``fit`` the fitted object returned from glm.
}
\description{
This estimates a distributed lag model (DLM) with a natrual spline basis.  The model is estimated with the glm function form the stats package.
}
