% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/drawpost.R
\name{drawpost}
\alias{drawpost}
\title{draw posterior for acpme}
\usage{
drawpost(number, y, X, C, int, scale, phi, nu, lambda)
}
\arguments{
\item{number}{Number of draws}

\item{y}{y}

\item{X}{X (or Z)}

\item{C}{C}

\item{int}{Indicator of an intercept}

\item{scale}{Scale for X.}

\item{phi}{phi}

\item{nu}{nu}

\item{lambda}{lambda}
}
\value{
posterior sample of beta
}
\description{
draw posterior for acpme
}

