% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hpd.R
\name{hpd}
\alias{hpd}
\title{Highest posterior density posterior intervals of a distribution}
\usage{
hpd(x, prob = 0.95)
}
\arguments{
\item{x}{Posterior sample}

\item{prob}{Probability to be contained in the interval.}
}
\value{
lower and upper bounds
}
\description{
Highest posterior density posterior intervals of a distribution
}

