% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.summary.acpme.R
\name{plot.acpme}
\alias{plot.acpme}
\title{Plot for acpme}
\usage{
\method{plot}{acpme}(x, ...)
}
\arguments{
\item{x}{An object of class 'acpme'.}

\item{...}{additional arguments for ggplot theme.}
}
\value{
message that there is no plot for an object of class acpme.
}
\description{
Plot for acpme
}
