% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bdlim.R
\name{plot.bdlim}
\alias{plot.bdlim}
\title{Plot for bdlim summary}
\usage{
\method{plot}{bdlim}(x, print = TRUE, ...)
}
\arguments{
\item{x}{An object of class 'bdlim'.}

\item{print}{A logical.  If TRUE then each plot will be printed. If FALSE then a list of plots will be returned.}

\item{...}{additional arguments for ggplot theme.}
}
\value{
If print=FALSE then an object of class ggplot is returned.
}
\description{
Plot for bdlim summary
}

