% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.dlm.R
\name{plot.dlm}
\alias{plot.dlm}
\title{Plot for dlm}
\usage{
\method{plot}{dlm}(x, ...)
}
\arguments{
\item{x}{An object of class 'dlm'.}

\item{...}{additional arguments for plot}
}
\description{
Plot for dlm
}
