% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.summary.acpme.R
\name{plot.summary.acpme}
\alias{plot.summary.acpme}
\title{Plot for acpme}
\usage{
\method{plot}{summary.acpme}(x, print = TRUE, ...)
}
\arguments{
\item{x}{An object of class 'summary.acpme'.}

\item{print}{A logical.  If TRUE then each plot will be printed. If FALSE then a list of plots will be returned.}

\item{...}{additional arguments for ggplot theme.}
}
\value{
If print=FALSE then a list of plots is returned.
}
\description{
Plot for acpme
}

