% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PMweekly-data.R
\docType{data}
\name{PMweekly}
\alias{PMweekly}
\title{Weekly average PM2.5 data}
\format{An object of class \code{"data.frame"}.}
\usage{
data(PMweekly)
}
\description{
Weekly average PM2.5 data for 1000 hypothetical births.
}
\examples{
data(PMweekly)
matplot(t(PMweekly[1:5,-c(1:2)]),type="l")
}
\keyword{datasets}

