% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.acpme.R
\name{print.acpme}
\alias{print.acpme}
\title{Default print for acpme object}
\usage{
\method{print}{acpme}(x, ...)
}
\arguments{
\item{x}{object of class acpme.}

\item{...}{additional arguments}
}
\description{
Default print for acpme object
}

