% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bdlim.R
\name{print.bdlim}
\alias{print.bdlim}
\title{Default print for bdlim object}
\usage{
\method{print}{bdlim}(x, ...)
}
\arguments{
\item{x}{bdlim.summary object to print}

\item{...}{additional arguments}
}
\description{
Default print for bdlim object
}
