% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bkmrdlm.R
\name{print.bkmrdlm}
\alias{print.bkmrdlm}
\title{Default print for summary.bkmrdlm object}
\usage{
\method{print}{bkmrdlm}(x, ...)
}
\arguments{
\item{x}{summary.bkmrdlm object to print}

\item{...}{additional arguments}
}
\description{
Default print for summary.bkmrdlm object
}
