% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.dlm.R
\name{print.dlm}
\alias{print.dlm}
\title{Default print for dlm object}
\usage{
\method{print}{dlm}(x, ...)
}
\arguments{
\item{x}{dlm object to print}

\item{...}{additional arguments}
}
\description{
Default print for dlm object
}
