% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwisebdlim.R
\name{print.pairwisebdlim}
\alias{print.pairwisebdlim}
\title{Default print for pairwise.bdlim object}
\usage{
\method{print}{pairwisebdlim}(x, ...)
}
\arguments{
\item{x}{pairwise.bdlim object to print}

\item{...}{additional arguments}
}
\description{
Default print for pairwise.bdlim object
}
