% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.acpme.R
\name{print.summary.acpme}
\alias{print.summary.acpme}
\title{Default print for bdlim.summary object}
\usage{
\method{print}{summary.acpme}(x, ...)
}
\arguments{
\item{x}{bdlim.summary object to print}

\item{...}{additional arguments}
}
\description{
Default print for bdlim.summary object
}

