% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.bdlim.R
\name{print.summary.bdlim}
\alias{print.summary.bdlim}
\title{Default print for bdlim.summary object}
\usage{
\method{print}{summary.bdlim}(x, ...)
}
\description{
Default print for bdlim.summary object
}

