% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bkmrdlm.R
\name{print.summary.bkmrdlm}
\alias{print.summary.bkmrdlm}
\title{Default print for summary.bkmrdlm object}
\usage{
\method{print}{summary.bkmrdlm}(x, ...)
}
\arguments{
\item{x}{summary.bkmrdlm object to print}

\item{...}{additional arguments}
}
\description{
Default print for summary.bkmrdlm object
}
