% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regimes-package.R
\docType{package}
\name{regimes}
\alias{regimes}
\title{Regression in Multivariate Exposure Settings}
\description{
Regression methods for estimating the relationship between a multivariate exposure and an outcome. The package is oriented toward environmental health and environmental epidemiology research but is applicable in a variety of settings. The three main areas that are addressed in the the package are: 1) exposure to multiple agents; 2) exposure to one or more agents over time (e.g. over the course of pregnancy); and 3) confounder adjustment when estimate the effect of a multivariate exposure.
}
\author{
ander wilson <ander.wilson@colostate.edu>
}
