% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simACPME.R
\name{simACPME}
\alias{simACPME}
\title{Simulate data for ACPME}
\usage{
simACPME(n, model = 1, p, m.main)
}
\arguments{
\item{n}{Integer number of observations.}

\item{model}{Takes values 1 (default) or 2 to indicate the choice of model to simulate data from.}

\item{p}{Number of covariates. Default is 100 and must be at least 35.}

\item{m.main}{Number of main effects for the exposure. Only used for model 2 and defaults to 2.}
}
\description{
This function simulated data used in the paper.
}

