% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simBDLIM.R
\name{simBDLIM}
\alias{simBDLIM}
\title{Simulate data from BDLIM}
\usage{
simBDLIM(n, design)
}
\arguments{
\item{n}{Number of observations must be between 100 and 1000}

\item{design}{Scenario from paper, simulation B, scenarios 1-5.}
}
\description{
This function uses a simular design to scenario 2 in the BDLIM paper but with different data.
}

