% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simregimes.R
\name{simregimes}
\alias{simregimes}
\title{Simulate data to demonstrate methods in regimes package.}
\usage{
simregimes(scenario, seed, ...)
}
\arguments{
\item{scenario}{Model to be simulated from. Options include: acpme1, and acpme2.}

\item{seed}{An optional random seed.}

\item{...}{Additional options for specific simulated data sets.}
}
\value{
A list containing the simulated data, true paramters, and opts which speficies the scenario being simulated from.
}
\description{
Simulate data to demonstrate methods in regimes package.
}
\examples{
dat_acpme1 <- simregimes(scenario="acpme1", seed=1234, n=200, p=100)
dat_acpme2 <- simregimes(scenario="acpme2", seed=1234, n=200, p=100, m.main=3)
}

