% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.acpme.R
\name{summary.acpme}
\alias{summary.acpme}
\title{Default summary for acpme}
\usage{
\method{summary}{acpme}(object, ciprob = 0.95, hpd.interval = FALSE,
  estimate, ...)
}
\arguments{
\item{object}{An object of class 'acpme'.}

\item{ciprob}{The probability contained by the posterior intervals.}

\item{hpd.interval}{Logical indicating if highest posterior density intervals should be computed (TRUE) or symmetric intervals (FALSE, default)}

\item{estimate}{A covtor or matrix defining a linear combination of the regression coeficients for the exposure.}

\item{...}{additional arguments}
}
\value{
A object of class 'summary.acpme'.
}
\description{
Default summary for acpme
}

