% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bkmrdlm.R
\name{summary.bkmrdlm}
\alias{summary.bkmrdlm}
\title{Default summary for bkmrdlm}
\usage{
\method{summary}{bkmrdlm}(object, alpha = 0.05, ...)
}
\arguments{
\item{object}{An object of class 'bkmrdlm'.}

\item{alpha}{The alpha level of significance to be used.}

\item{...}{Additional arguments affecting the summary produced.}
}
\value{
A object of class 'summary.bkmrdlm'
}
\description{
Default summary for bkmrdlm
}
