% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdlimbasis.R
\name{bdlimbasis}
\alias{bdlimbasis}
\title{Build BDLIM basis}
\usage{
bdlimbasis(X, knots = NULL, df = NULL, pve = 0.99, type = "face")
}
\arguments{
\item{X}{A n x p matrix of observations. Each row has the observations for one individual. The observations should be evenly spaced over a grid.}

\item{knots}{The number of knots to use or a vector of knots.}

\item{df}{Degrees of freedom for the natrual spline basis.}

\item{pve}{Proportion of variance explained by the PC basis using FACE. Must be in (0,1].}

\item{type}{The type of basis expansion to use.  Default is 'face' corresponding the the FACE method. Alternative method is 'ns' for a natrual cubic spline basis.}
}
\value{
A list with elements:
\item{psi}{Matrix with eigenfunctions in the columns.}
\item{lambda}{Vector of eigen values.}
\item{knots}{Knots used (if applicable).}
\item{df}{Degrees of freedom used (if applicable).}
\item{type}{The type of basis constructed.}
}
\description{
\code{bdlimbasis} builds the basis for thge BDLIM model using either FACE of natural cubic splines.
}
\author{
Ander Wilson
}
\seealso{
\code{\link{fpca.face}} and \code{\link{ns}} for details on the methods called.
}

