% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdlimglmall.R
\name{bdlimglmall}
\alias{bdlimglmall}
\title{BDLIM for glm}
\usage{
bdlimglmall(Y, X, Z, G, B, model, niter, nburn, nthin, prior, post)
}
\arguments{
\item{Y}{Outcome vector}

\item{X}{Exposure matrix}

\item{Z}{Matrix of covariates. An intercept will be added}

\item{G}{Vector indicating group membership}

\item{B}{Basis object from build.basis}

\item{model}{Indicator of how G enters the model. If 'bw' both the betas and weight functions will vary between groups. If 'b' then only the betas will vary between groups. If 'w' then only the wieght function will vary by group. If "n" then neither beta or the wieght function will vary by group. If a vector of any combination of the four choices is provided then those models will be run and compared. The option 'all' runs all four models and compares them.}

\item{niter}{Number of MCMC iterations}

\item{nburn}{Number of MCMC iterations to be discarded as burning}

\item{nthin}{Number of draws taken to obtain one sample}

\item{prior}{List with the entries:  betavar = the prior variance for beta; and gamma = the prior variance for the covarites. The priors on beta and gamma are iid normal mean zero.}

\item{post}{Relating to the posterior likelihood
#' @author Ander Wilson}
}
\description{
This estimates the model for a single group or the overall effect
}

