% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdlimlmoverall.R
\name{bdlimlmoverall}
\alias{bdlimlmoverall}
\title{bdlimlmoverall}
\usage{
bdlimlmoverall(Y, X, Z, G, B, niter, nburn, nthin, prior)
}
\arguments{
\item{Y}{Outcome vector}

\item{X}{Exposure matrix}

\item{Z}{Matrix of covariates. An intercept will be added}

\item{G}{Vector indicating group membership}

\item{B}{Basis object from build.basis}

\item{niter}{Number of MCMC iterations}

\item{nburn}{Number of MCMC iterations to be discarded as burning}

\item{nthin}{Number of draws taken to obtainone sampl}

\item{prior}{List with the entries: sigma = a numeric 2-vector with the shape and rate paramters for the pirior in the error precision (1/sigma^2); betavar = the prior variance for beta; and gamma = the prior variance for the covarites. The priors on beta and gamma are iid normal mean zero.}
}
\description{
This estimates the model for a single group or the overall effect
}
\author{
Ander Wilson
}

