% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/beta.bdlim.R
\name{beta.bdlim}
\alias{beta.bdlim}
\title{Summary of the beta for BDLIM}
\usage{
\method{beta}{bdlim}(fit, inter.model, ciprob = 0.95, hpd.interval = FALSE)
}
\arguments{
\item{fit}{An object of class 'bdlim'.}

\item{inter.model}{Model to be summarized.  The default is \code{inter.model}=1 indicating to summarize the best fitting model.  \code{inter.model}=2, 3, or 4 indicates to summarized the second, third, or fourth best fitting model respectively. Model fit is determined by posterior probability. Alternative, 'BDLIM_n', 'BDLIM_bw', 'BDLIM_b', or 'BDLIM_w' can be entered to return a specific model.}

\item{ciprob}{The probability contained by the posterior intervals.}

\item{hpd.interval}{Logical indicating if highest posterior density intervals should be computed (TRUE) or symmetric intervals (FALSE, default)}
}
\value{
Data.frame summarizing the posterior distribution.
}
\description{
Summary of the beta for BDLIM
}

