% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawpost.R
\name{drawpost}
\alias{drawpost}
\title{Draw posterior samples for acpme}
\usage{
drawpost(number, y, X, C, int, scale, phi, nu, lambda)
}
\arguments{
\item{number}{number of samples}

\item{y}{y}

\item{X}{X or Z}

\item{C}{C}

\item{int}{Logical indicating there is an intercept}

\item{scale}{scale for X}

\item{phi}{phi}

\item{nu}{nu}

\item{lambda}{lambda}
}
\value{
A posterior sample of regression coeficiences for X
}
\description{
Draw posterior samples for acpme
}
