% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makepen.R
\name{makepen}
\alias{makepen}
\title{Make penalty}
\usage{
makepen(Z, C, pen.type = "eigen")
}
\arguments{
\item{Z}{Matrix of exposures including interactions.}

\item{C}{A n x p matrix of covaraites.}

\item{pen.type}{Choice of penalty. The default is "eigen." Other options are "correlation" and "projection."}
}
\description{
This function computes the penalty weights (omega) for each covaraite.
}
\examples{
dat_acpme1 <- simregimes(scenario="acpme1", seed=1234, n=200, p=100)
omega <- makepen(Z=scale(dat_acpme1$Z),C=scale(dat_acpme1$C))
par(mfrow=c(2,1))
plot(omega$omega)
pr <- exp(omega$omega*omega$lambda)/(1+exp(omega$omega*omega$lambda))
plot(pr)
}

