% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bkmrdlm.R
\name{plot.bkmrdlm}
\alias{plot.bkmrdlm}
\title{Plot functions for BKMR-DLM}
\usage{
\method{plot}{bkmrdlm}(x, print = TRUE, ...)
}
\arguments{
\item{x}{An object of class 'bkmrdlm'.}

\item{print}{A logical.  If TRUE then each plot will be printed. If FALSE then a list of plots will be returned.}

\item{...}{Additional arguments for ggplot theme.}
}
\value{
If print=FALSE then an object of class ggplot is returned.
}
\description{
Plot functions for BKMR-DLM
}
