% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.summary.bdlim.R
\name{plot.summary.bdlim}
\alias{plot.summary.bdlim}
\title{Plot for bdlim}
\usage{
\method{plot}{summary.bdlim}(x, print = TRUE)
}
\arguments{
\item{x}{An object of class 'summary.bdlim'.}

\item{print}{A logical.  If TRUE then each plot will be printed. If FALSE then a list of plots will be returned.}
}
\value{
If print=FALSE then a list of plots is returned.
}
\description{
Plot for bdlim
}

