% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.summary.bdlim.R
\name{plot.summary.bdlim}
\alias{plot.summary.bdlim}
\title{Plot for bdlim summary}
\usage{
\method{plot}{summary.bdlim}(x, print = TRUE, blackwhite = FALSE, ...)
}
\arguments{
\item{x}{An object of class 'summary.bdlim'.}

\item{print}{A logical.  If TRUE then each plot will be printed. If FALSE then a list of plots will be returned.}

\item{blackwhite}{A logical.  If TRUE the credible regrion ribbons are grey. If FALSE (default) then they are different colors for each group.}

\item{...}{additional arguments for ggplot theme.}
}
\value{
If print=FALSE then a list of plots is returned.
}
\description{
Plot for bdlim summary
}

