% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictUnivariate.R
\name{predictUnivariate}
\alias{predictUnivariate}
\title{Predict univariate result for polynomial kernel BKMR-DLM}
\usage{
predictUnivariate(object, points = 20, crossM, qtl = 0.5)
}
\arguments{
\item{object}{An object of class bkmrdlm.}

\item{points}{Number of points to predict at}

\item{crossM}{Exposure to set a cross section quantile for}

\item{qtl}{quantile for the cross section}
}
\value{
a dataframe containing predicted values.
}
\description{
Predict univariate result for polynomial kernel BKMR-DLM
}
