% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.acpme.R
\name{print.summary.acpme}
\alias{print.summary.acpme}
\title{Default print for acpme.summary object}
\usage{
\method{print}{summary.acpme}(x, minpr = 0.5, ...)
}
\arguments{
\item{x}{acpme.summary object to print}

\item{minpr}{The confounder inclusion probabilities for covariates with posterior inclusion probability at least minpr will be printed.}

\item{...}{additional arguments}
}
\description{
Default print for acpme.summary object
}
