% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bdlim.R
\name{print.summary.bdlim}
\alias{print.summary.bdlim}
\title{Default print for summary.bdlim object}
\usage{
\method{print}{summary.bdlim}(x, ...)
}
\arguments{
\item{x}{summary.bdlim object to print}

\item{...}{additional arguments}
}
\description{
Default print for summary.bdlim object
}
