% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simBKMRDLM.R
\name{simBKMRDLM}
\alias{simBKMRDLM}
\title{Simulate data for BKMR-DLM.}
\usage{
simBKMRDLM(n = 200, scenario = "A", sd = 1, seed, dta)
}
\arguments{
\item{n}{Number of observations.}

\item{scenario}{Simulation scenario.}

\item{sd}{Error standard deviation.}

\item{seed}{An optional seed.}

\item{dta}{A list with data to use. The elements should be X, a list of exposure matrices, and Z a matrix of covariates. If not data is provided then data will be simulated automatically.}
}
\value{
A list containing simulated data
}
\description{
Simulate data for BKMR-DLM.
}
