% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.bdlim.R
\name{summary.bdlim}
\alias{summary.bdlim}
\title{summary for bdlim}
\usage{
\method{summary}{bdlim}(object, inter.model, ciprob, hpd.interval, ...)
}
\arguments{
\item{object}{An object of class 'bdlim'.}

\item{inter.model}{Model to be summarized.  The default is \code{inter.model}=1 indicating to summarize the best fitting model.  \code{inter.model}=2, 3, or 4 indicates to summarized the second, third, or fourth best fitting model respectively. Model fit is determined by posterior probability. Alternative, 'BDLIM_n', 'BDLIM_bw', 'BDLIM_b', or 'BDLIM_w' can be entered to return a specific model.}

\item{ciprob}{The probability contained by the posterior intervals.}

\item{hpd.interval}{Logical indicating if highest posterior density intervals should be computed (TRUE) or symmetric intervals (FALSE, default)}

\item{...}{other things to be specificed.}
}
\value{
An object of class 'summary.bdlim'.
}
\description{
summary for bdlim
}

